chore(validation): Separate types and validation code
